Hybrid_Variant/requirements.txt
torch
torchvision
numpy
scikit-learn
matplotlib
seaborn
opencv-python
xgboost
catboost
pandas
tqdm
Pillow
glob2